moved alering service and alert class to shared dir
