+++ src/app/components/register/RegisterComponent.spec.ts
@@ -9,19 +9,19 @@ import {FormBuilder} from 'angular2/common';
 import {Router} from 'angular2/router';
 import {Observable} from 'rxjs/Rx';
 
+import {RegisterComponent} from './RegisterComponent';
 import {User} from '../../shared/models/User';
 import {ImagesService} from '../../shared/services/ImagesService';
 import {GlobalService} from '../../shared/services/GlobalService';
 import {UserService} from '../../shared/services/UserService';
 import {UserValidationService} from '../../shared/services/UserValidationService';
-import {AlertingService} from '../alerting/AlertingService';
-import {RegisterComponent} from './RegisterComponent';
-import {Alert} from '../alerting/Alert';
+import {AlertingService} from '../../shared/services/AlertingService';
+import {Alert} from '../../shared/models/Alert';
+import {ValidationResponse} from '../../shared/models/ValidationResponse';
 
 import {UserServiceMock} from '../../shared/mocks/UserServiceMock';
 import {RouterMock} from '../../shared/mocks/RouterMock';
 import {ImagesServiceMock} from '../../shared/mocks/ImagesServiceMock';
-import {ValidationResponse} from '../../shared/models/ValidationResponse';
 
 describe('RegisterComponentTests', () => {
   beforeEachProviders(() => [
