+++ src/app/components/upload/UploadPictureComponent.spec.ts
@@ -8,7 +8,7 @@ import {provide} from 'angular2/core';
 import {UploadPictureComponent} from './UploadPictureComponent';
 import {UploadPictureService} from './UploadPictureService';
 import {GlobalService} from '../../shared/services/GlobalService';
-import {AlertingService} from '../alerting/AlertingService';
+import {AlertingService} from '../../shared/services/AlertingService';
 
 describe('UploadPictureComponentTests', () => {
   beforeEachProviders(() => [
