+++ src/app/components/upload/UploadPictureService.ts
@@ -2,7 +2,7 @@ import {Injectable, bind} from 'angular2/core';
 import {Http} from 'angular2/http';
 
 import {GlobalService} from '../../shared/services/GlobalService';
-import {AlertingService} from '../alerting/AlertingService';
+import {AlertingService} from '../../shared/services/AlertingService';
 import {MultipartItem} from '../../shared/plugins/multipart-upload/multipart-item';
 import {MultipartUploader} from '../../shared/plugins/multipart-upload/multipart-uploader';
