+++ src/app/shared/services/services.ts
@@ -4,7 +4,7 @@ import {userSettingsServiceInjectables} from './UserSettingsService';
 import {imagesServiceInjectables} from './ImagesService';
 import {userValidationServiceInjectables} from './UserValidationService';
 import {uploadPictureServiceInjectables} from '../../components/upload/UploadPictureService';
-import {alertingServiceInjectables} from '../../components/alerting/AlertingService';
+import {alertingServiceInjectables} from '../../shared/services/AlertingService';
 import {userSettingsColorsServiceInjectables} from '../../components/userSettings/UserSettingsColorsService';
 
 export * from './GlobalService';
@@ -12,8 +12,8 @@ export * from './UserService';
 export * from './UserSettingsService';
 export * from './ImagesService';
 export * from './UserValidationService';
+export * from './AlertingService';
 export * from '../../components/upload/UploadPictureService';
-export * from '../../components/alerting/AlertingService';
 export * from '../../components/userSettings/UserSettingsColorsService';
 
 export var servicesInjectables: Array<any> = [
